Refactor admin dashboard into modules
Refactors the admin dashboard page into smaller, more manageable modules for improved code organization and maintainability.
